Fixed Clickprototype
Graph-QL in Appv1

Starting Vue for homepage
