+++ eborrow-app/html/start_page.html
@@ -4,12 +4,13 @@
 <head>
     <meta charset="UTF-8">
     <title>Game-Tracker</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="../css/style.css">
+    <link rel="stylesheet" href="../css/startpage_style.css"> 
 </head>
 
 <body>
     <header>
-        <img class="logo" src="images/logo.svg" alt="logo">
+        <img class="logo" src="../images/logo.svg" alt="logo">
         <nav class="navigation">
             <ul class="nav-links">
                 <li><a href="#">Home</a></li>
@@ -44,27 +45,27 @@
         <div class="overview">
             <table class="overview-table-container">
                 <tr class="overview-entry">
-                    <td class="overview-image"><img src="images/dummy-image.jpg" alt="logo"></td>
+                    <td class="overview-image"><img src="../images/dummy-image.jpg" alt="logo"></td>
                     <td class="overview-description">Hier steht was zum Spiel</td>
                 </tr>
                 <tr class="overview-entry">
-                    <td class="overview-image"><img src="images/dummy-image.jpg" alt="logo"></td>
+                    <td class="overview-image"><img src="../images/dummy-image.jpg" alt="logo"></td>
                     <td class="overview-description">Hier steht was zum Spiel</td>
                 </tr>
                 <tr class="overview-entry">
-                    <td class="overview-image"><img src="images/dummy-image.jpg" alt="logo"></td>
+                    <td class="overview-image"><img src="../images/dummy-image.jpg" alt="logo"></td>
                     <td class="overview-description">Hier steht was zum Spiel</td>
                 </tr>
                 <tr class="overview-entry">
-                    <td class="overview-image"><img src="images/dummy-image.jpg" alt="logo"></td>
+                    <td class="overview-image"><img src="../images/dummy-image.jpg" alt="logo"></td>
                     <td class="overview-description">Hier steht was zum Spiel</td>
                 </tr>
                 <tr class="overview-entry">
-                    <td class="overview-image"><img src="images/dummy-image.jpg" alt="logo"></td>
+                    <td class="overview-image"><img src="../images/dummy-image.jpg" alt="logo"></td>
                     <td class="overview-description">Hier steht was zum Spiel</td>
                 </tr>
                 <tr class="overview-entry">
-                    <td class="overview-image"><img src="images/dummy-image.jpg" alt="logo"></td>
+                    <td class="overview-image"><img src="../images/dummy-image.jpg" alt="logo"></td>
                     <td class="overview-description">Hier steht was zum Spiel</td>
                 </tr>
             </table>
