+++ eborrow-frontend/app-v2/package.json
@@ -23,7 +23,7 @@
     "eslint": "^9.14.0",
     "eslint-plugin-vue": "^9.30.0",
     "prettier": "^3.3.3",
-    "vite": "^6.0.1",
+    "vite": "^6.0.6",
     "vite-plugin-vue-devtools": "^7.6.5"
   }
 }
