+++ eborrow-frontend/app-v2/src/composables/useGamesRestApi.js
@@ -1,12 +1,9 @@
-const allGamesEndpoint = "http://localhost:8080/home";
+'use strict'
 
-const allGenresEndpoint = "http://localhost:8080/game/all-genres";
-const allPlatformsEndpoint = "http://localhost:8080/game/all-platforms";
+const allGamesEndpoint = "http://localhost:8080/home";
 
 const gamesBySearchQueryURL = "http://localhost:8080/game/games-search";
 const gamesByFilterParamsURL = "http://localhost:8080/game/filtered-games";
-
-const graphQLURL = "http:///localhost:8080/graphql";
 
 function handleGameFilling(games) {
     const table = document.querySelector(".overview-table-container");
@@ -99,80 +96,13 @@ async function getGamesByFilter() {
     developerInput.value = "";
 }
 
-async function requestAllGamesGraphQL(){
-    let result;
+async function initSortBar(){
     
-    const body = JSON.stringify({
-        query: `{
-          games {
-                description
-                title
-                gameImage
-            }
-        }`,
-    });
-    
-    try{
-        const myHeaders = new Headers();
-        myHeaders.append("Content-Type", "application/json");
-    
-        const response = await fetch(graphQLURL, {
-            method: "POST",
-            body: body,
-            headers: myHeaders,
-        });
-        result = await response.json();
-        result = result["data"]["games"];
-    }catch (error) {
-        console.error(error.message)
-    }
-    return result;
 }
 
-async function startUp() {
-
-    const games = await requestAllGamesGraphQL();
+async function initGamesOverview() {
+    const games = await requestResource(allGamesEndpoint);
     handleGameFilling(games);
-
-    const genres = await requestResource(allGenresEndpoint);
-    const genreSelection = document.querySelector(".genre-filter");
-    genreSelection.innerHTML = '';
-
-    const allGenres = document.createElement("option");
-    allGenres.setAttribute("value", "All");
-    allGenres.appendChild(document.createTextNode("All"));
-    genreSelection.appendChild(allGenres);
-
-    for (var i = 0; i < genres.length; i++) {
-        const option = document.createElement("option");
-        option.setAttribute("value", genres[i]);
-        const optionTextNode = document.createTextNode(genres[i]);
-        option.appendChild(optionTextNode);
-        genreSelection.appendChild(option);
-    }
-
-    const platforms = await requestResource(allPlatformsEndpoint);
-    const platformsSelection = document.querySelector(".platform-filter");
-    platformsSelection.innerHTML = '';
-
-    const allPlatforms = document.createElement("option");
-    allPlatforms.setAttribute("value", "All");
-    allPlatforms.appendChild(document.createTextNode("All"));
-    platformsSelection.appendChild(allPlatforms);
-
-    for (var i = 0; i < platforms.length; i++) {
-        const option = document.createElement("option");
-        option.setAttribute("value", platforms[i]);
-        const optionTextNode = document.createTextNode(platforms[i]);
-        option.appendChild(optionTextNode);
-        platformsSelection.appendChild(option);
-    }
-
-    const searchButton = document.querySelector(".searchButton");
-    searchButton.addEventListener("click", getGamesBySearchQuery);
-
-    const filterButton = document.querySelector(".apply-filter");
-    filterButton.addEventListener("click", getGamesByFilter);
 }
 
-startUp();
+export {initGamesOverview, getGamesByFilter, getGamesBySearchQuery, requestResource}
